[#71] Remove ambiguity around uAttributes requirements
The requirements regarding the usage of particular UAttributes
properties for the message types supported by uProtocol had been
scattered around multiple documents, some textual descriptions as well
as proto files. It was hard to find the relevant information and to
keep it consistent across the specifications.

The following things have been done to mitigate the situation:
* The uattributes.adoc file is now the single source of truth regarding
  the usage of UAttributes for the supported message types. All
  corresponding definitions have been removed from the cloudevents.adoc
  and uattributes.proto files.
* The cloudevents.adoc file has been moved to the up-l1 folder because
  its content is relevant to uProtocol Client libraries only that use
  CloudEvents' JSON or Protobuf Format to map UMessages to a transport
  protocol's PDU. Application (uEntity) developers will not get in
  touch with CloudEvents at all.