Removing another --file, fixing the proto root path
The --file flag is not needed since pom.xml is in the repo root now.

The proto source root was also pointing to the wrong place now that the
xml file has been moved up a directory.

Also fixing the URLs for the project/repository since they were pointing
to the old core API archive.