Service instance id clarification
The following change adds some verbiage that states that it is valid to not set the instance_id that indicates that we do not care. This is often used when referring to services that do not have more than one instance.

This change is to help clarify the requirements after looking at the C++ UUri validators.

#217