Service instance id clarification
The following changes the wildcard for ue instance id to be 0xFFFF in lieu of 0x0000 that was meant to refer to the default instance when there is only a single instance of a service. Separating the default instance from the wildcard makes for simpler pattern matchers. Also fixed a few minor issues in the spec verbiage.