Adding OFT metadata and uTransport address spoofing check.
The transport needs to verify that the caller (client) source address is the same as what is in the passed `UMessage` of the send() API to avoid any uE from spoofing another uE.  The only exception for this rule is if the uE is the streamer who is simply forwarding events from one transport to the next and in that case the streamers identity needs to be verified upon creation of the transport.

The PR also adds in missing OFT metadata.

#218