Remove unused SubscriberInfo from some USubscription messages
The SubscriberInfo was a wrapper for the subscribers URI however it passes details that was never used (and complicates the implementations) and is passed in some scenarios that is redundant (we know who is calling the APIs). The SubscriberInfo is also passed to registerForNotifications however subscribers don't call this API so it is confusing for developers.