Make use of existing MQTT 5 header properties
Adapted the mapping of UAttribute fields to make use of standard
properties defined in MQTT 5.

Also removed the distinct spec item identifiers defined for each field.
The mapping of the fields will most likely happen in very coherent
code blocks (functions) so it should be sufficient to trace the
requirements at that level.