Adapt specitem references to reflect recent changes
One of the specitems in the uProtocol URI spec has been updated and
assigned a new revision number.

The Gherkin scenarios in the uuri_uri_serialization.feature file
refer to the old revision number. This commit updates those references
to the new revision number to ensure that the scenarios accurately reflect
the current state of the specification.